feat: Add debug log mask support (#40)
* feat: Add debug log mask support

NGINX supports multiple debug masks to customize logging.

We use Rust's type system to represent these as an enum and ease a
developer's usage. They can call the ngx_log_debug_mask with any of the
supported enum elements.

To aid in keeping in sync with NGINX primitives a unit test will fail
when the current FIRST and LAST log masks do not align any longer.