refactor(log): reimplement with traits on objects containing ngx_log_t
A new `LogTarget` trait can be implemented for any object that owns a
logger or requires a special log behavior.

The logger implementation can do a mut cast on the `ngx_log_t` pointer
to avoid borrowing the target mutably.  This looks highly suspicious,
but is a defined behavior and mostly safe.