refactor: rename `NGX_OBJS` to `NGINX_BUILD_DIR`
The `NGX_OBJS` name was taken from the nginx buildsystem, but it is an
internal detail there, so the name wasn't really meaningful.  Now that
we have `NGINX_SOURCE_DIR`, there's a benefit in more consistent naming.

Technically, not a breaking change, as we haven't had a release with the
old variable name.