feat(sys): reimplement `ngx_queue_t` macros
Bindgen does not support functional macros or inline functions, so we
have to maintain a copy of the implementation of any such item we need
to use.

There are no safe wrappers, because all the operations here are unsafe
and unsound, unless used on the heap- or pool-allocated objects with
certain ownership constraints.